Book generation and implemented turns
The book manager now generates books at the start of the game and at the
start of every new turn. Number of books is randomized to show that it
works properly, but a fixed amount can be used later.

Each turn is split into a picking phase and a browsing phase. The book
manager is only visible and accessible during the picking phase, after
which it is hidden until the start of the next turn. The game manager
checks if each player is done with the current phase before moving on to
the next. Pressing Y on the controller indicates that you are ready to
begin a new turn.