Worked on book generation and control issues
Books will now be generated with different colors and patterns. The
texture manager takes care of loading and preparing the textures from
the correct folder so adding more colors and patterns does not require
any extra code. Added a selection square instead of simply tinting the
books to help with visibility with the different colored books.

Also fixed some bugs with regards to the controller input that popped up
after the last merge.